backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyPDF2==3.0.1
python-dotenv==1.0.0
supabase==2.1.0
google-generativeai==0.3.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24.0,<0.25.0